refactored

+++ FileUploadAndValidation/Models/NasBillPaymentDto.cs
@@ -6,10 +6,12 @@ namespace FileUploadAndValidation.Models
 {
     public class NasBillPaymentDto
     {
-        public double amount { get; set; }
+        public decimal amount { get; set; }
         public string customer_id { get; set; }
         public int row { get; set; }
         public string item_code { get; set; }
         public string product_code { get; set; }
     }
+
+
 }
